+++ chapter9-5-1/src/main/java/com/goat/Z/README.md
@@ -1,3 +1,6 @@
+# 目录
+    item01 和 item02 演示了 依赖倒置原则
+
 # 里氏替换原则
     1. 定义：里氏替换原则简单易懂一点的定义就是只要父类出现的地方子类就可以出现，且替换成子类也不会出现任何错误或者异常。
     (但是反过来，有子类出现的地方，父类不一定可以适用)。
