+++ chapter9-5-1/src/main/java/com/goat/Z/Z06/example04/Driver1.java
@@ -1,6 +1,6 @@
 package com.goat.Z.Z06.example04;
 
-import com.goat.Z.Z06.example03.Car;
+import com.goat.Z.Z06.item02.Car;
 
 /**
  * Created by 64274 on 2019/5/6.
@@ -10,7 +10,9 @@ import com.goat.Z.Z06.example03.Car;
  * @ date 2019/5/6---13:42
  */
 public class Driver1 {
+
     private Car car;
+
     Driver1(Car car){
         this.car = car;
     }
