+++ chapter9-5-1/src/main/java/com/goat/Z/Z06/item01/App.java
@@ -1,8 +1,10 @@
-package com.goat.Z.Z06.example01;
+package com.goat.Z.Z06.item01;
+
+import org.junit.Test;
 
 /**
  * Created by 64274 on 2019/5/6.
- *
+ * 实例类之间产生依赖所出现的问题
  * @ Description: 假如有一天Driver不开benz了，则此时代码要修改两处：
  *  1.Driver类中的参数 public void drive(Benz benz)
  *  2.App类中的调用参数  driver.drive(new BMW());
@@ -13,8 +15,12 @@ package com.goat.Z.Z06.example01;
 
 public class App {
 
-    public static void main(String[] args){
-        Driver driver = new Driver();
-        driver.drive(new Bmw()); // 此处需要修改
+    Driver driver = new Driver();
+
+    @Test
+    public void test(){
+        driver.drive(new Bmw()); //
+//        driver.drive(new Benz());// 开奔驰则报错  需要修改
     }
+
 }
