+++ chapter9-5-1/src/main/java/com/goat/Z/Z06/item02/Benz.java
@@ -1,4 +1,4 @@
-package com.goat.Z.Z06.example03;
+package com.goat.Z.Z06.item02;
 
 /**
  * Created by 64274 on 2019/5/6.
